+++ NEWS.md
@@ -1,3 +1,6 @@
 # nlmixr2targets 0.0.0.9000
 
+* Setting initial conditions can now be done with `cmt(initial)` which will be
+  automatically recoded to `cmt(0)`.
 * Added "cens" and "limit" columns as nlmixr2 columns
+* Keep columns from the `table` argument's keep element